requirements.txt
flask==3.0.0
requests==2.31.0
openai==1.46.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
httpx==0.27.2

